+++ DeepLearningShogi.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "train", "train\train.pyproj", "{44DC35CD-F353-4DA5-AC49-F1A723D9ABF6}"
+Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "DeepLearningShogi", "DeepLearningShogi.pyproj", "{B1194544-1E7D-4F6E-8C32-FC6887DCDFF5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,8 +11,8 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{44DC35CD-F353-4DA5-AC49-F1A723D9ABF6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{44DC35CD-F353-4DA5-AC49-F1A723D9ABF6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B1194544-1E7D-4F6E-8C32-FC6887DCDFF5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B1194544-1E7D-4F6E-8C32-FC6887DCDFF5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
