+++ bat/search1_player.bat
@@ -1,2 +1,3 @@
 @echo off
-python -m pydlshogi.usi.usi_search1_player
+call activate DeepLearningShogi
+python -m dlshogi.usi.usi_search1_player
