+++ dlshogi/usi/usi_mcts_player.py
@@ -1,0 +1,5 @@
+﻿from dlshogi.usi.usi import *
+from dlshogi.player.mcts_player import *
+
+player = MCTSPlayer()
+usi(player)
