+++ dlshogi/usi/usi_parallel_mcts_player.py
@@ -1,0 +1,5 @@
+﻿from dlshogi.usi.usi import *
+from dlshogi.player.parallel_mcts_player import *
+
+player = ParallelMCTSPlayer()
+usi(player)
