+++ dlshogi/usi/usi_policy_player.py
@@ -1,0 +1,5 @@
+﻿from dlshogi.usi.usi import *
+from dlshogi.player.policy_player import *
+
+player = PolicyPlayer()
+usi(player)
