+++ dlshogi/usi/usi_search1_player.py
@@ -1,0 +1,5 @@
+﻿from dlshogi.usi.usi import *
+from dlshogi.player.search1_player import *
+
+player = Search1Player()
+usi(player)
