--- pydlshogi/usi/usi_mcts_player.py
@@ -1,5 +1,0 @@
-﻿from pydlshogi.usi.usi import *
-from pydlshogi.player.mcts_player import *
-
-player = MCTSPlayer()
-usi(player)
