--- pydlshogi/usi/usi_parallel_mcts_player.py
@@ -1,5 +1,0 @@
-﻿from pydlshogi.usi.usi import *
-from pydlshogi.player.parallel_mcts_player import *
-
-player = ParallelMCTSPlayer()
-usi(player)
