--- pydlshogi/usi/usi_policy_player.py
@@ -1,5 +1,0 @@
-﻿from pydlshogi.usi.usi import *
-from pydlshogi.player.policy_player import *
-
-player = PolicyPlayer()
-usi(player)
