--- pydlshogi/usi/usi_search1_player.py
@@ -1,5 +1,0 @@
-﻿from pydlshogi.usi.usi import *
-from pydlshogi.player.search1_player import *
-
-player = Search1Player()
-usi(player)
