+++ setup.py
@@ -1,0 +1,9 @@
+﻿import setuptools
+
+setuptools.setup(
+    name = 'dlshogi',
+    version = '0.0.1',
+    author = 'test',
+    packages = ['dlshogi'],
+    scripts = [],
+)
